refactor: remove quantization subdirectory
