+++ quanto/qtensor/func.py
