+++ test/qtensor/ops/test_quantized_dispatch.py
@@ -2,7 +2,7 @@ import pytest
 import torch
 from helpers import q_assert_close, random_qtensor, random_tensor
 
-from quanto.quantization import QTensor
+from quanto import QTensor
 
 
 def test_to_device(device):
